fix: version null
