+++ app/src/main/java/its/madruga/warevamp/app/core/XposedChecker.java
@@ -1,4 +1,4 @@
-package its.madruga.warevamp.core;
+package its.madruga.warevamp.app.core;
 
 import de.robv.android.xposed.XC_MethodReplacement;
 import de.robv.android.xposed.XposedHelpers;
@@ -9,7 +9,7 @@ public class XposedChecker {
     }
 
     public static void setActiveModule(ClassLoader loader) {
-        XposedHelpers.findAndHookMethod("its.madruga.warevamp.core.XposedChecker",
+        XposedHelpers.findAndHookMethod("its.madruga.warevamp.app.core.XposedChecker",
                 loader,
                 "isActive",
                 XC_MethodReplacement.returnConstant(true));
