+++ app/src/main/java/its/madruga/warevamp/broadcast/Receivers.java
@@ -1,0 +1,11 @@
+package its.madruga.warevamp.broadcast;
+
+import its.madruga.warevamp.broadcast.receivers.ModuleReceiver;
+import its.madruga.warevamp.broadcast.receivers.WhatsAppReceiver;
+
+public class Receivers {
+    public static void registerReceivers() {
+        new ModuleReceiver().registerAllReceivers();
+        new WhatsAppReceiver().registerAllReceivers();
+    }
+}
