+++ app/src/main/java/its/madruga/warevamp/broadcast/receivers/ModuleReceiver.java
@@ -1,4 +1,6 @@
-package its.madruga.warevamp.core.broadcast.receivers;
+package its.madruga.warevamp.broadcast.receivers;
+
+import android.util.Log;
 
 import its.madruga.warevamp.App;
 
@@ -13,6 +15,5 @@ public class ModuleReceiver extends EventReceiver {
 
     @Override
     public void registerAllReceivers() {
-
     }
 }
